refactor: di layer local data base 주입 로직 변경
