+++ android/app/src/main/java/com/zzang/chongdae/ChongdaeApp.kt
@@ -1,11 +1,8 @@
 package com.zzang.chongdae
 
 import android.app.Application
-import android.content.Context
-import androidx.datastore.preferences.preferencesDataStore
 import com.google.firebase.FirebaseApp
 import com.kakao.sdk.common.KakaoSdk
-import com.zzang.chongdae.data.local.database.AppDatabase
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
@@ -14,16 +11,5 @@ class ChongdaeApp : Application() {
         super.onCreate()
         KakaoSdk.init(this, BuildConfig.NATIVE_APP_KEY)
         FirebaseApp.initializeApp(this)
-        _chongdaeAppContext = this
-    }
-
-    companion object {
-        val Context.dataStore by preferencesDataStore(name = "member_preferences")
-
-        private lateinit var _chongdaeAppContext: Context
-        val chongdaeAppContext get() = _chongdaeAppContext
-
-        private val appDatabase: AppDatabase by lazy { AppDatabase.getInstance(chongdaeAppContext) }
-        val offeringDao by lazy { appDatabase.offeringDao() }
     }
 }
