+++ android/di/src/main/kotlin/com/zzang/chongdae/di/core/DataStoreExt.kt
@@ -1,0 +1,10 @@
+package com.zzang.chongdae.di.core
+
+import android.content.Context
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.preferencesDataStore
+
+private const val DATA_STORE_NAME = "user_preferences"
+
+val Context.dataStore: DataStore<Preferences> by preferencesDataStore(name = DATA_STORE_NAME)
