+++ android/di/src/main/kotlin/com/zzang/chongdae/di/core/DataStoreModule.kt
@@ -1,19 +1,14 @@
-package com.zzang.chongdae.di.module
+package com.zzang.chongdae.di.core
 
 import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.preferencesDataStore
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
-
-private const val DATA_STORE_NAME = "user_preferences"
-
-private val Context.dataStore: DataStore<Preferences> by preferencesDataStore(name = DATA_STORE_NAME)
 
 @Module
 @InstallIn(SingletonComponent::class)
